Replaced Simple MediaEntryProvider with a Simple SessionProvider.
This approach is more versatile, and allows other components to use the simple session provider.
